202504/0426/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.7
matplotlib>=3.5.0
Pillow>=8.0.0 